+++ index.d.ts
@@ -1,1 +1,4 @@
-export * from './dist/ng-select';
+export * from './dist/option.interface';
+export * from './dist/select.component';
+export declare class SelectModule {
+}