+++ src/ng-select.ts
@@ -5,6 +5,9 @@ import {FormsModule} from '@angular/forms';
 import {SelectComponent} from './select.component';
 import {SelectDropdownComponent} from './select-dropdown.component';
 
+export * from './option.interface';
+export * from './select.component';
+
 @NgModule({
     declarations: [
         SelectComponent,