+++ src/option.interface.ts
@@ -1,0 +1,5 @@
+export interface IOption {
+    value: string;
+    label: string;
+    disabled?: boolean;
+}